Get rid of circular deps in batch
