+++ src/views/batch/BatchView.test.tsx
@@ -1,12 +1,11 @@
 import { OpKind, OperationContentsAndResult } from "@taquito/rpc";
 import { TezosOperationError } from "@taquito/taquito";
 
-import { BatchView, tokenTitle } from "./BatchView";
+import { BatchView } from "./BatchView";
 import { mockImplicitAccount, mockTezOperation } from "../../mocks/factories";
 import { addAccount, mockEstimatedFee } from "../../mocks/helpers";
 import { act, render, screen, userEvent, within } from "../../mocks/testUtils";
 import { mockToast } from "../../mocks/toast";
-import { ghostFA2, ghostTezzard } from "../../mocks/tokens";
 import { makeAccountOperations } from "../../types/AccountOperations";
 import { Operation } from "../../types/Operation";
 import { estimate } from "../../utils/tezos";
@@ -144,29 +143,5 @@ describe("<BatchView />", () => {
       expect(estimationStatuses[1]).toHaveTextContent(/^Estimated$/);
       expect(estimationStatuses[2]).toHaveTextContent(/^Estimated$/);
     });
-  });
-});
-
-describe("tokenTitle", () => {
-  it("returns raw amount if token is missing", () => {
-    expect(tokenTitle(undefined, "1000000")).toBe("1000000 Unknown Token");
-  });
-
-  it("doesn't return symbol if token name is absent", () => {
-    expect(tokenTitle(ghostTezzard, "1000000")).toBe("1000000 Tezzardz #24");
-  });
-
-  it("returns symbol if name is absent", () => {
-    const token = ghostTezzard;
-    delete token.metadata.name;
-    expect(tokenTitle(token, "1000000")).toBe("1000000 FKR");
-  });
-
-  it("returns just amount if neither name nor symbol are present", () => {
-    expect(tokenTitle({ ...ghostTezzard, metadata: {} }, "1000000")).toBe("1000000");
-  });
-
-  it("returns pretty amount if decimals field is present", () => {
-    expect(tokenTitle(ghostFA2, "1000321")).toBe("10.00321 Klondike3");
   });
 });
