+++ src/views/batch/tokenTitle.ts
@@ -1,0 +1,16 @@
+import { compact } from "lodash";
+
+import { Token, tokenName, tokenPrettyAmount, tokenSymbol } from "../../types/Token";
+
+export const tokenTitle = (token: Token | undefined, amount: string) => {
+  if (!token) {
+    return `${amount} Unknown Token`;
+  }
+  const name = tokenName(token);
+  const prettyAmount = tokenPrettyAmount(amount, token, { showSymbol: false });
+
+  // don't show the symbol if the token name is present
+  const symbol = name ? undefined : tokenSymbol(token);
+
+  return compact([prettyAmount, symbol, name]).join(" ");
+};
